feat-8: added Drawer,when we have display less desktop
