+++ src/components/Drawer/Drawer.tsx
@@ -1,0 +1,2 @@
+import { Drawer } from "@mui/material";
+export default Drawer;
