+++ src/components/index.ts
@@ -5,3 +5,4 @@ export { TabPanel } from "./TabPanel";
 export { Typography } from "./Typography";
 export { Grid } from "./Grid";
 export { AppBar } from "./AppBar";
+export { Drawer } from "./Drawer";
